# openapi.yaml
openapi: 3.0.0
info:
  title: Faculty Reporting System API
  version: 1.0.0
  description: API documentation for the Faculty Reporting System (backend in Express.js)
servers:
  - url: http://localhost:5003/api
    description: Local development server (client .env default)
  - url: http://localhost:5000/api
    description: Local development server (server default)
  - url: http://localhost:5001/api
    description: Local development server (alternate)
security:
  - BearerAuth: []
paths:
...
  /reports:
    get:
      summary: Get all reports
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Create a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  report:
                    $ref: '#/components/schemas/Report'
...
  /users:
    get:
      summary: Get all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        is_active:
          type: boolean
    UserCreateRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role:
          type: string
          enum: [student, lecturer, program_leader, principal_lecturer, faculty_manager, admin]
        phone: { type: string }
...
    Report:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        report_type: { type: string }
        status: { type: string }
        rating: { type: integer, nullable: true }
        attendance_count: { type: integer, nullable: true }
        topic_covered: { type: string, nullable: true }
        learning_outcomes: { type: string, nullable: true }
        challenges: { type: string, nullable: true }
        recommendations: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        reporter_role: { type: string, nullable: true }
        course_name: { type: string, nullable: true }
        course_code: { type: string, nullable: true }
    ReportCreateRequest:
      type: object
      required: [title, content, reportType]
      properties:
        title: { type: string }
        content: { type: string }
        reportType: { type: string }
        courseId: { type: integer, nullable: true }
        attendanceCount: { type: integer, nullable: true }
        topicCovered: { type: string, nullable: true }
        learningOutcomes: { type: string, nullable: true }
        challenges: { type: string, nullable: true }
        recommendations: { type: string, nullable: true }
        rating: { type: integer, nullable: true }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...